Fix bugs

+++ database/migrations/2020_07_15_203945_create_customers_table.php
@@ -16,7 +16,7 @@ class CreateCustomersTable extends Migration
             $table->string('bill');
             $table->decimal('pay')->nullable();
             $table->string('financing');
-            $table->decimal('adders')->nullable();
+            $table->integer('adders')->nullable();
             $table->decimal('epc')->nullable();
             $table->decimal('commission')->nullable();
             $table->decimal('setter_fee')->nullable();
